% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwgraph.R
\name{idwgraph}
\alias{idwgraph}
\title{Network IDW Interpolation}
\usage{
idwgraph(roads, aprs, power = 2)
}
\arguments{
\item{roads}{An sf object with LINESTRING geometry (the road network).}

\item{aprs}{An sf object with POINT geometry and a 'value' column (the observations).}

\item{power}{The IDW power parameter (default = 2).}
}
\value{
An sf object of the network nodes with interpolated values.
}
\description{
Network IDW Interpolation
}
